Feat: Survival b4 & Stricter seed production
Mortality is the first process to happen. There is no more selective mortality of individuals that did not grow and the current production is calculated inside that function.
Seed production is even more strict: adults are explicitly filtered and the number of offsprings has to be positive, even if they are only produced by individuals with positive reproductive biomass.

Related issues: #641